# Conflicts:
#	LAPR5_3DD_019/Migrations/MDFContextModelSnapshot.cs
#	LAPR5_3DD_019/bin/Debug/netcoreapp2.2/LAPR5_3DD_019.dll
#	LAPR5_3DD_019/bin/Debug/netcoreapp2.2/LAPR5_3DD_019.pdb
#	LAPR5_3DD_019/bin/Debug/netcoreapp2.2/LAPR5_3DD_019.runtimeconfig.dev.json
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/LAPR5_3DD_019.assets.cache
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/LAPR5_3DD_019.csproj.CoreCompileInputs.cache
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/LAPR5_3DD_019.csprojAssemblyReference.cache
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/LAPR5_3DD_019.dll
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/LAPR5_3DD_019.pdb
#	LAPR5_3DD_019/obj/Debug/netcoreapp2.2/project.razor.json
#	LAPR5_3DD_019/obj/LAPR5_3DD_019.csproj.nuget.cache
#	LAPR5_3DD_019/obj/LAPR5_3DD_019.csproj.nuget.dgspec.json
#	LAPR5_3DD_019/obj/LAPR5_3DD_019.csproj.nuget.g.props
#	LAPR5_3DD_019/obj/project.assets.json

+++ LAPR5_3DD_019/Models/DTO/TipoMaquinaDTO.cs
@@ -7,12 +7,11 @@ namespace LAPR5_3DD_019.Models.DTO
     public class TipoMaquinaDTO
     {
 
-        public float Id_tipoMaquina { get; set; }
+        public long Id_tipoMaquina { get; set; }
         public string descricaoTipoMaquina { get; set; }
-
         public ICollection<OperacaoDTO> operacoes { get; set; }
 
-        public TipoMaquinaDTO(float Id, string descricaoTipoMaquina, ICollection<TipoMaquinaOperacao> operacoes)
+        public TipoMaquinaDTO(long Id, string descricaoTipoMaquina, ICollection<TipoMaquinaOperacao> operacoes)
         {
             this.Id_tipoMaquina = Id;
             this.descricaoTipoMaquina = descricaoTipoMaquina;
